Add initial e2e test/setup
