+++ self_signed_test.go
@@ -1,3 +1,5 @@
+// +build !e2e
+
 /*
 Copyright 2018 All rights reserved.
