+++ server_test.go
@@ -1,3 +1,5 @@
+// +build !e2e
+
 /*
 Copyright 2015 All rights reserved.
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,102 +18,15 @@ limitations under the License.
 package main
 
 import (
-	"crypto/tls"
-	"encoding/json"
-	"fmt"
 	"net/http"
 	"net/http/httptest"
-	"net/url"
 	"strings"
 	"testing"
 	"time"
 
-	"crypto/x509"
-	"encoding/pem"
-
 	"github.com/stretchr/testify/assert"
-	"golang.org/x/net/websocket"
-	gojose "gopkg.in/square/go-jose.v2"
 	"gopkg.in/square/go-jose.v2/jwt"
 )
-
-const (
-	fakeAdminRole          = "role:admin"
-	fakeAdminRoleURL       = "/admin*"
-	fakeAuthAllURL         = "/auth_all/*"
-	fakeClientID           = "test"
-	fakeSecret             = "test"
-	fakeTestAdminRolesURL  = "/test_admin_roles"
-	fakeTestRole           = "role:test"
-	fakeTestRoleURL        = "/test_role"
-	fakeTestWhitelistedURL = "/auth_all/white_listed*"
-	testProxyAccepted      = "Proxy-Accepted"
-	validUsername          = "test"
-	validPassword          = "test"
-)
-
-type RoleClaim struct {
-	Roles []string `json:"roles"`
-}
-
-type DefaultTestTokenClaims struct {
-	Aud               string               `json:"aud"`
-	Azp               string               `json:"azp"`
-	ClientSession     string               `json:"client_session"`
-	Email             string               `json:"email"`
-	FamilyName        string               `json:"family_name"`
-	GivenName         string               `json:"given_name"`
-	Username          string               `json:"username"`
-	Iat               int64                `json:"iat"`
-	Iss               string               `json:"iss"`
-	Jti               string               `json:"jti"`
-	Name              string               `json:"name"`
-	Nbf               int                  `json:"nbf"`
-	Exp               int64                `json:"exp"`
-	PreferredUsername string               `json:"preferred_username"`
-	SessionState      string               `json:"session_state"`
-	Sub               string               `json:"sub"`
-	Typ               string               `json:"typ"`
-	Groups            []string             `json:"groups"`
-	RealmAccess       RoleClaim            `json:"realm_access"`
-	ResourceAccess    map[string]RoleClaim `json:"resource_access"`
-	Item              string               `json:"item"`
-	Found             string               `json:"found"`
-	Item1             []string             `json:"item1"`
-	Item2             []string             `json:"item2"`
-	Item3             []string             `json:"item3"`
-}
-
-var defTestTokenClaims = DefaultTestTokenClaims{
-	Aud:               "test",
-	Azp:               "clientid",
-	ClientSession:     "f0105893-369a-46bc-9661-ad8c747b1a69",
-	Email:             "gambol99@gmail.com",
-	FamilyName:        "Jayawardene",
-	GivenName:         "Rohith",
-	Username:          "Jayawardene",
-	Iat:               1450372669,
-	Iss:               "test",
-	Jti:               "4ee75b8e-3ee6-4382-92d4-3390b4b4937b",
-	Name:              "Rohith Jayawardene",
-	Nbf:               0,
-	Exp:               0,
-	PreferredUsername: "rjayawardene",
-	SessionState:      "98f4c3d2-1b8c-4932-b8c4-92ec0ea7e195",
-	Sub:               "1e11e539-8256-4b3b-bda8-cc0d56cddb48",
-	Typ:               "Bearer",
-	Groups:            []string{"default"},
-	RealmAccess:       RoleClaim{Roles: []string{"default"}},
-	ResourceAccess: map[string]RoleClaim{
-		"defaultclient": {
-			Roles: []string{"default"},
-		},
-	},
-	Item:  "item",
-	Item1: []string{"default"},
-	Item2: []string{"default"},
-	Item3: []string{"default"},
-}
 
 func TestNewKeycloakProxy(t *testing.T) {
 	cfg := newFakeKeycloakConfig()
@@ -967,315 +882,4 @@ func TestDisableAuthorizationCookie(t *testing.T) {
 		},
 	}
 	p.RunTests(t, requests)
-}
-
-func newTestService() string {
-	_, _, u := newTestProxyService(nil)
-	return u
-}
-
-func newTestProxyService(config *Config) (*oauthProxy, *fakeAuthServer, string) {
-	if config == nil {
-		config = newFakeKeycloakConfig()
-	}
-
-	authConfig := &fakeAuthConfig{}
-	if config.SkipOpenIDProviderTLSVerify {
-		authConfig.EnableTLS = true
-	}
-
-	auth := newFakeAuthServer(authConfig)
-
-	config.DiscoveryURL = auth.getLocation()
-	config.RevocationEndpoint = auth.getRevocationURL()
-	config.Verbose = false
-	config.EnableLogging = false
-
-	proxy, err := newProxy(config)
-	if err != nil {
-		panic("failed to create proxy service, error: " + err.Error())
-	}
-
-	// step: create an fake upstream endpoint
-	proxy.upstream = new(fakeUpstreamService)
-	service := httptest.NewServer(proxy.router)
-	config.RedirectionURL = service.URL
-
-	// step: we need to update the client config
-	if proxy.provider, proxy.idpClient, err = proxy.newOpenIDProvider(); err != nil {
-		panic("failed to recreate the openid client, error: " + err.Error())
-	}
-
-	return proxy, auth, service.URL
-}
-
-func newFakeHTTPRequest(method, path string) *http.Request {
-	return &http.Request{
-		Method: method,
-		Header: make(map[string][]string),
-		Host:   "127.0.0.1",
-		URL: &url.URL{
-			Scheme: "http",
-			Host:   "127.0.0.1",
-			Path:   path,
-		},
-	}
-}
-
-func newFakeKeycloakConfig() *Config {
-	return &Config{
-		ClientID:                    fakeClientID,
-		ClientSecret:                fakeSecret,
-		CookieAccessName:            "kc-access",
-		CookieRefreshName:           "kc-state",
-		DisableAllLogging:           true,
-		DiscoveryURL:                "127.0.0.1:0",
-		EnableAuthorizationCookies:  true,
-		EnableAuthorizationHeader:   true,
-		EnableLogging:               false,
-		EnableLoginHandler:          true,
-		EnableTokenHeader:           true,
-		EnableCompression:           false,
-		EnableMetrics:               false,
-		Listen:                      "127.0.0.1:0",
-		ListenAdmin:                 "",
-		ListenAdminScheme:           "http",
-		TLSAdminCertificate:         "",
-		TLSAdminPrivateKey:          "",
-		TLSAdminCaCertificate:       "",
-		OAuthURI:                    "/oauth",
-		OpenIDProviderTimeout:       time.Second * 5,
-		SkipOpenIDProviderTLSVerify: false,
-		SkipUpstreamTLSVerify:       false,
-		Scopes:                      []string{},
-		Verbose:                     false,
-		Resources: []*Resource{
-			{
-				URL:     fakeAdminRoleURL,
-				Methods: []string{"GET"},
-				Roles:   []string{fakeAdminRole},
-			},
-			{
-				URL:     fakeTestRoleURL,
-				Methods: []string{"GET"},
-				Roles:   []string{fakeTestRole},
-			},
-			{
-				URL:     fakeTestAdminRolesURL,
-				Methods: []string{"GET"},
-				Roles:   []string{fakeAdminRole, fakeTestRole},
-			},
-			{
-				URL:     fakeAuthAllURL,
-				Methods: allHTTPMethods,
-				Roles:   []string{},
-			},
-			{
-				URL:         fakeTestWhitelistedURL,
-				WhiteListed: true,
-				Methods:     allHTTPMethods,
-				Roles:       []string{},
-			},
-		},
-	}
-}
-
-func makeTestCodeFlowLogin(location string) (*http.Response, []*http.Cookie, error) {
-	flowCookies := make([]*http.Cookie, 0)
-
-	u, err := url.Parse(location)
-
-	if err != nil {
-		return nil, nil, err
-	}
-	// step: get the redirect
-	var resp *http.Response
-	for count := 0; count < 4; count++ {
-		req, err := http.NewRequest(http.MethodGet, location, nil)
-
-		if err != nil {
-			return nil, nil, err
-		}
-
-		if resp != nil {
-			cookies := resp.Cookies()
-			flowCookies = append(flowCookies, cookies...)
-		}
-
-		// step: make the request
-		tr := &http.Transport{
-			TLSClientConfig: &tls.Config{
-				//nolint:gas
-				InsecureSkipVerify: true,
-			},
-		}
-
-		resp, err = tr.RoundTrip(req)
-
-		if err != nil {
-			return nil, nil, err
-		}
-
-		if resp.StatusCode != http.StatusSeeOther {
-			return nil, nil, fmt.Errorf("no redirection found in resp, status code %d", resp.StatusCode)
-		}
-
-		location = resp.Header.Get("Location")
-
-		if !strings.HasPrefix(location, "http") && !strings.HasPrefix(location, "https") {
-			location = fmt.Sprintf("%s://%s%s", u.Scheme, u.Host, location)
-		}
-	}
-	return resp, flowCookies, nil
-}
-
-// fakeUpstreamResponse is the response from fake upstream
-type fakeUpstreamResponse struct {
-	URI     string      `json:"uri"`
-	Method  string      `json:"method"`
-	Address string      `json:"address"`
-	Headers http.Header `json:"headers"`
-}
-
-// fakeUpstreamService acts as a fake upstream service, returns the headers and request
-type fakeUpstreamService struct{}
-
-func (f *fakeUpstreamService) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set(testProxyAccepted, "true")
-
-	upgrade := strings.ToLower(r.Header.Get("Upgrade"))
-	if upgrade == "websocket" {
-		websocket.Handler(func(ws *websocket.Conn) {
-			defer ws.Close()
-			var data []byte
-			err := websocket.Message.Receive(ws, &data)
-			if err != nil {
-				ws.WriteClose(http.StatusBadRequest)
-				return
-			}
-			content, _ := json.Marshal(&fakeUpstreamResponse{
-				URI:     r.RequestURI,
-				Method:  r.Method,
-				Address: r.RemoteAddr,
-				Headers: r.Header,
-			})
-			_ = websocket.Message.Send(ws, content)
-		}).ServeHTTP(w, r)
-	} else {
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusOK)
-		content, _ := json.Marshal(&fakeUpstreamResponse{
-			// r.RequestURI is what was received by the proxy.
-			// r.URL.String() is what is actually sent to the upstream service.
-			// KEYCLOAK-10864, KEYCLOAK-11276, KEYCLOAK-13315
-			URI:     r.URL.String(),
-			Method:  r.Method,
-			Address: r.RemoteAddr,
-			Headers: r.Header,
-		})
-		_, _ = w.Write(content)
-	}
-}
-
-type fakeToken struct {
-	claims DefaultTestTokenClaims
-}
-
-func newTestToken(issuer string) *fakeToken {
-	claims := defTestTokenClaims
-	claims.Exp = time.Now().Add(1 * time.Hour).Unix()
-	claims.Iat = time.Now().Unix()
-	claims.Iss = issuer
-
-	return &fakeToken{claims: claims}
-}
-
-// getToken returns a JWT token from the clains
-func (t *fakeToken) getToken() (string, error) {
-	input := []byte("")
-	block, _ := pem.Decode([]byte(fakePrivateKey))
-	if block != nil {
-		input = block.Bytes
-	}
-
-	var priv interface{}
-	priv, err0 := x509.ParsePKCS8PrivateKey(input)
-
-	if err0 != nil {
-		return "", err0
-	}
-
-	alg := gojose.SignatureAlgorithm("RS256")
-	privKey := &gojose.JSONWebKey{Key: priv, Algorithm: string(alg), KeyID: "test-kid"}
-	signer, err := gojose.NewSigner(gojose.SigningKey{Algorithm: alg, Key: privKey}, nil)
-
-	if err != nil {
-		return "", err
-	}
-
-	b := jwt.Signed(signer).Claims(&t.claims)
-	jwt, err := b.CompactSerialize()
-
-	if err != nil {
-		return "", err
-	}
-
-	return jwt, nil
-}
-
-// getUnsignedToken returns a unsigned JWT token from the clains
-func (t *fakeToken) getUnsignedToken() (string, error) {
-	input := []byte("")
-	block, _ := pem.Decode([]byte(fakePrivateKey))
-	if block != nil {
-		input = block.Bytes
-	}
-
-	var priv interface{}
-	priv, err0 := x509.ParsePKCS8PrivateKey(input)
-
-	if err0 != nil {
-		return "", err0
-	}
-
-	alg := gojose.SignatureAlgorithm("RS256")
-	privKey := &gojose.JSONWebKey{Key: priv, Algorithm: string(alg), KeyID: ""}
-	signer, err := gojose.NewSigner(gojose.SigningKey{Algorithm: alg, Key: privKey}, nil)
-
-	if err != nil {
-		return "", err
-	}
-
-	b := jwt.Signed(signer).Claims(&t.claims)
-	jwt, err := b.CompactSerialize()
-
-	if err != nil {
-		return "", err
-	}
-
-	items := strings.Split(jwt, ".")
-	jwt = strings.Join(items[0:1], ".")
-
-	return jwt, nil
-}
-
-// setExpiration sets the expiration of the token
-func (t *fakeToken) setExpiration(tm time.Time) {
-	t.claims.Exp = tm.Unix()
-}
-
-// addGroups adds groups to then token
-func (t *fakeToken) addGroups(groups []string) {
-	t.claims.Groups = groups
-}
-
-// addRealmRoles adds realms roles to token
-func (t *fakeToken) addRealmRoles(roles []string) {
-	t.claims.RealmAccess.Roles = roles
-}
-
-// addClientRoles adds client roles to the token
-func (t *fakeToken) addClientRoles(client string, roles []string) {
-	t.claims.ResourceAccess = make(map[string]RoleClaim)
-	t.claims.ResourceAccess[client] = RoleClaim{Roles: roles}
 }
